Fix for cubestate being over written by GUI request
When a solution was underway, with the calibration window open and contours on,
the gui was continually polling for frames. This also initiated a solve if the
cubestate list was empty.

Hence, when the 'actual' solution button was pressed, the cubestate was emptied
to accept the updated cubestate. However, during this time, the GUI would also
poll, see that the cubelist was empty and attempt another solution.
Obviously, this completely messes with things